+++ examples/bizyair_box_point_guided_segment-anything.json
@@ -1,0 +1,262 @@
+{
+  "last_node_id": 21,
+  "last_link_id": 18,
+  "nodes": [
+    {
+      "id": 3,
+      "type": "PreviewImage",
+      "pos": {
+        "0": 1574,
+        "1": 392,
+        "2": 0,
+        "3": 0,
+        "4": 0,
+        "5": 0,
+        "6": 0,
+        "7": 0,
+        "8": 0,
+        "9": 0
+      },
+      "size": {
+        "0": 210,
+        "1": 246
+      },
+      "flags": {},
+      "order": 2,
+      "mode": 0,
+      "inputs": [
+        {
+          "name": "images",
+          "type": "IMAGE",
+          "link": 17
+        }
+      ],
+      "outputs": [],
+      "properties": {
+        "Node name for S&R": "PreviewImage"
+      }
+    },
+    {
+      "id": 12,
+      "type": "PreviewImage",
+      "pos": {
+        "0": 1855,
+        "1": 389,
+        "2": 0,
+        "3": 0,
+        "4": 0,
+        "5": 0,
+        "6": 0,
+        "7": 0,
+        "8": 0,
+        "9": 0
+      },
+      "size": {
+        "0": 210,
+        "1": 246
+      },
+      "flags": {},
+      "order": 4,
+      "mode": 0,
+      "inputs": [
+        {
+          "name": "images",
+          "type": "IMAGE",
+          "link": 10
+        }
+      ],
+      "outputs": [],
+      "properties": {
+        "Node name for S&R": "PreviewImage"
+      }
+    },
+    {
+      "id": 11,
+      "type": "MaskToImage",
+      "pos": {
+        "0": 1810,
+        "1": 310,
+        "2": 0,
+        "3": 0,
+        "4": 0,
+        "5": 0,
+        "6": 0,
+        "7": 0,
+        "8": 0,
+        "9": 0
+      },
+      "size": {
+        "0": 264.5999755859375,
+        "1": 26
+      },
+      "flags": {},
+      "order": 3,
+      "mode": 0,
+      "inputs": [
+        {
+          "name": "mask",
+          "type": "MASK",
+          "link": 18
+        }
+      ],
+      "outputs": [
+        {
+          "name": "IMAGE",
+          "type": "IMAGE",
+          "links": [
+            10
+          ],
+          "slot_index": 0,
+          "shape": 3
+        }
+      ],
+      "properties": {
+        "Node name for S&R": "MaskToImage"
+      }
+    },
+    {
+      "id": 19,
+      "type": "Note",
+      "pos": {
+        "0": 1132,
+        "1": 160,
+        "2": 0,
+        "3": 0,
+        "4": 0,
+        "5": 0,
+        "6": 0,
+        "7": 0,
+        "8": 0,
+        "9": 0
+      },
+      "size": {
+        "0": 289.005615234375,
+        "1": 115.42251586914062
+      },
+      "flags": {},
+      "order": 0,
+      "mode": 0,
+      "inputs": [],
+      "outputs": [],
+      "properties": {},
+      "widgets_values": [
+        "SAM EDITOR功能：\n\n可选功能：point模式和box模式\n\n使用步骤：☁️BizyAir Point-Box Guided SAM 节点上右键选择Open in SAM EDITOR，可以通过Select mode按钮切换模式，point模式下通过鼠标左键单击选择抠图区域，box模式下拖动鼠标进行框选\n\nPoint mode按钮： 工作在point mode下，point mode按钮精准控制分割区域，点中区域是想保留得区域（remain）或者要去除得区域（remove）\n\nClear按钮：清除画布上得所有选择抠图区域\n\nUndo按钮：撤销上次选择得抠图区域\n\nCancel：清除画布上得所有选择抠图区域并退出\n\nSave to node: 保存\n\n"
+      ],
+      "color": "#432",
+      "bgcolor": "#653"
+    },
+    {
+      "id": 20,
+      "type": "BizyAirSegmentAnythingPointBox",
+      "pos": {
+        "0": 1124,
+        "1": 375,
+        "2": 0,
+        "3": 0,
+        "4": 0,
+        "5": 0,
+        "6": 0,
+        "7": 0,
+        "8": 0,
+        "9": 0
+      },
+      "size": {
+        "0": 307.80706787109375,
+        "1": 338
+      },
+      "flags": {},
+      "order": 1,
+      "mode": 0,
+      "inputs": [],
+      "outputs": [
+        {
+          "name": "IMAGE",
+          "type": "IMAGE",
+          "links": [
+            17
+          ],
+          "slot_index": 0,
+          "shape": 3
+        },
+        {
+          "name": "MASK",
+          "type": "MASK",
+          "links": [
+            18
+          ],
+          "slot_index": 1,
+          "shape": 3
+        }
+      ],
+      "properties": {
+        "Node name for S&R": "BizyAirSegmentAnythingPointBox"
+      },
+      "widgets_values": [
+        "0804.png",
+        true,
+        "image"
+      ]
+    }
+  ],
+  "links": [
+    [
+      10,
+      11,
+      0,
+      12,
+      0,
+      "IMAGE"
+    ],
+    [
+      17,
+      20,
+      0,
+      3,
+      0,
+      "IMAGE"
+    ],
+    [
+      18,
+      20,
+      1,
+      11,
+      0,
+      "MASK"
+    ]
+  ],
+  "groups": [
+    {
+      "title": "选择抠图区域",
+      "bounding": [
+        1114,
+        309,
+        351,
+        410
+      ],
+      "color": "#A88",
+      "font_size": 24
+    },
+    {
+      "title": "抠图结果",
+      "bounding": [
+        1490,
+        308,
+        642,
+        341
+      ],
+      "color": "#b58b2a",
+      "font_size": 24
+    }
+  ],
+  "config": {},
+  "extra": {
+    "ds": {
+      "scale": 1.3310000000000006,
+      "offset": [
+        -800.1960907672773,
+        15.658618943562546
+      ]
+    }
+  },
+  "version": 0.4
+}
